Update um_setup.rst
I have tried with the specified command on the installation instructions to create the Cate environment under Linux, and it did not work.
I am new to Python, so it has taken me a while to figure out what was wrong.
By installing anaconda or miniconda, and trying to create the environment as described in this document, it basically goes to Anaconda Cloud. I am not sure if this is due to the version of Miniconda that I installed, but actually it is the one listed at this document too.
Basically I have researched a bit and verified that the environment creation is with the following syntax: "conda create --name cate-env".
With that, I then activated with "conda activate cate-env" (for Linux it is stated as "source" rather than "conda", but it does work as requested by the inner conda command after creating the environment). "Source" is actually to load in the current shell the new environment. I just closed and reopen the terminal shell to apply the changes.